Moved helper code out of __init__.py
Preparing for the terse executable loader
